+++ app/controllers/assignments_controller.rb
@@ -349,6 +349,7 @@ class AssignmentsController < ApplicationController
     summary = {
       name: "#{assignment.short_identifier}: #{assignment.description}",
       average: assignment.results_average(points: true) || 0,
+      average_annotations: assignment.average_annotations,
       median: assignment.results_median(points: true) || 0,
       max_mark: assignment.max_mark || 0,
       standard_deviation: assignment.results_standard_deviation || 0,